Working tank health bar and updates

+++ server/src/main/java/edu/unh/cs/cs619/bulletzone/repository/GameRepository.java
@@ -28,7 +28,7 @@ public interface GameRepository {
     boolean ejectPowerUp(long playableId)
             throws TankDoesNotExistException;
 
-    Long getLife(long playableId, int playableType)
+    int getLife(long playableId, int playableType)
         throws TankDoesNotExistException;
 
     public void leave(long playableId)
